+++ baiducloud-python-sdk-rapidfs/baiducloud_python_sdk_rapidfs/models/auth_info.py
@@ -14,10 +14,10 @@ class AuthInfo(AbstractModel):
         """
         Initialize AuthInfo instance.
 
-        :param cidr: 权限地址，ip/ip 段，请确保格式正确有效，如：127.0.0.1，127.0.0.0/28
+        :param cidr: 权限地址，IP/IP 段，请确保格式正确有效，如：127.0.0.1，127.0.0.0/28
         :type cidr: str (optional)
 
-        :param mode: 权限类型，枚举值：ReadOnly：只读；ReadWrite：可读写；Forbid：禁止访问
+        :param mode: 权限类型，枚举值：* ReadOnly：只读；* ReadWrite：可读写；* Forbid：禁止访问
         :type mode: str (optional)
 
         :param description: 权限描述信息