+++ baiducloud-python-sdk-rapidfs/baiducloud_python_sdk_rapidfs/models/check_before_create_instance_response.py
@@ -11,18 +11,18 @@ class CheckBeforeCreateInstanceResponse(BceResponse):
     CheckBeforeCreateInstanceResponse
     """
 
-    def __init__(self, var_pass=None, err_info=None):
+    def __init__(self, rapidfs_pass=None, err_info=None):
         """
         Initialize CheckBeforeCreateInstanceResponse response.
 
-        :param var_pass: 是否通过校验
-        :type var_pass: bool (optional)
+        :param rapidfs_pass: 是否通过校验
+        :type rapidfs_pass: bool (optional)
 
         :param err_info: err_info field
         :type err_info: ErrInfo (optional)
         """
         super().__init__()
-        self.var_pass = var_pass
+        self.rapidfs_pass = rapidfs_pass
         self.err_info = err_info
 
     def to_dict(self):
@@ -41,8 +41,8 @@ class CheckBeforeCreateInstanceResponse(BceResponse):
         result = dict()
         if self.metadata is not None:
             result['metadata'] = dict(self.metadata)
-        if self.var_pass is not None:
-            result['pass'] = self.var_pass
+        if self.rapidfs_pass is not None:
+            result['pass'] = self.rapidfs_pass
         if self.err_info is not None:
             result['errInfo'] = self.err_info.to_dict()
         return result
@@ -64,7 +64,7 @@ class CheckBeforeCreateInstanceResponse(BceResponse):
         """
         m = m or dict()
         if m.get('pass') is not None:
-            self.var_pass = m.get('pass')
+            self.rapidfs_pass = m.get('pass')
         if m.get('errInfo') is not None:
             self.err_info = ErrInfo().from_dict(m.get('errInfo'))
         return self