+++ baiducloud-python-sdk-rapidfs/ChangeLog.md
@@ -1,3 +1,7 @@
+2026-06-15 Version: 0.0.3
+- 涉及产品: RAPIDFS，元数据同步规则列表及详情接口字段IntervalMinutes更正为intervalMinutes
+- 涉及产品: RAPIDFS，缓存管理规则任务列表字段cacheRuleJobInfos和cacheRuleJobId更正为cacheJobInfos和cacheJobId
+
 2026-06-03 Version: 0.0.2
 - 涉及产品: RAPIDFS，元数据同步、缓存规则、实例管理等接口统一移除query.action参数